refactor: MemberListElement를 통한 코드 리팩토링
